+++ sample/src/main/java/com/github/panpf/assemblyadapter/sample/ui/recycler/RecyclerGridDividerVerFragment.kt
@@ -29,10 +29,7 @@ import androidx.navigation.fragment.navArgs
 import androidx.paging.LoadState
 import androidx.recyclerview.widget.ConcatAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.github.panpf.assemblyadapter.recycler.AssemblyGridLayoutManager
-import com.github.panpf.assemblyadapter.recycler.AssemblyRecyclerAdapter
-import com.github.panpf.assemblyadapter.recycler.AssemblySingleDataRecyclerAdapter
-import com.github.panpf.assemblyadapter.recycler.ItemSpan
+import com.github.panpf.assemblyadapter.recycler.*
 import com.github.panpf.assemblyadapter.recycler.divider.Divider
 import com.github.panpf.assemblyadapter.recycler.divider.Insets
 import com.github.panpf.assemblyadapter.recycler.divider.newAssemblyGridDividerItemDecoration
@@ -106,16 +103,14 @@ class RecyclerGridDividerVerFragment : ToolbarFragment<FragmentRecyclerDividerVe
             dividerParamsViewMode.dividerParamsData.observe(viewLifecycleOwner) { dividerParams ->
                 dividerParams ?: return@observe
 
-                layoutManager = AssemblyGridLayoutManager(
-                    requireContext(),
-                    dividerParams.getSpanCount(true),
-                    dividerParams.spanSizeByPosition,
-                    mapOf(
+                layoutManager = newAssemblyGridLayoutManager(dividerParams.getSpanCount(true)) {
+                    itemSpanByPosition(dividerParams.spanSizeByPosition ?: emptyMap())
+                    itemSpanByItemFactory(
                         AppsOverviewItemFactory::class to ItemSpan.fullSpan(),
                         ListSeparatorItemFactory::class to ItemSpan.fullSpan(),
                         LoadStateItemFactory::class to ItemSpan.fullSpan()
                     )
-                )
+                }
 
                 if (itemDecorationCount > 0) {
                     removeItemDecorationAt(0)